--- src/backend/backend.rs
@@ -1,4 +1,0 @@
-pub trait SigningBackend {
-    fn sign(&self, data: &[u8]) -> Vec<u8>;
-    fn public_key(&self) -> Vec<u8>;
-}
