+++ src/error.rs
@@ -39,25 +39,25 @@ pub enum SignerError {
     InvalidTimestamp,
 }
 impl From<base64::DecodeError> for SignerError {
-    fn from(b64_error: base64::DecodeError) -> SignerError {
+    fn from(_b64_error: base64::DecodeError) -> SignerError {
         SignerError::TODO // possibly invalid data? but i'd like to convey in the error what kind of data did we wanna decode
     }
 }
 
 impl From<std::array::TryFromSliceError> for SignerError {
-    fn from(slice_error: std::array::TryFromSliceError) -> SignerError {
+    fn from(_slice_error: std::array::TryFromSliceError) -> SignerError {
         SignerError::TODO // something like invalid key
     }
 }
 
 impl From<ed25519_consensus::Error> for SignerError {
-    fn from(ed25519_error: ed25519_consensus::Error) -> SignerError {
+    fn from(_ed25519_error: ed25519_consensus::Error) -> SignerError {
         SignerError::TODO // something like invalid key... too?
     }
 }
 
 impl From<toml::de::Error> for SignerError {
-    fn from(toml_error: toml::de::Error) -> SignerError {
+    fn from(_toml_error: toml::de::Error) -> SignerError {
         SignerError::InvalidConfig
     }
 }
